+++ src/components/modal/ModalDronerInfinity.tsx
@@ -5,7 +5,7 @@ import FooterPage from '../footer/FooterPage'
 import { useNavigate } from 'react-router-dom'
 import { DeleteOutlined, StarFilled } from '@ant-design/icons'
 import { color, icon } from '../../resource'
-import upload_droner_infinity from '../../resource/media/empties/upload_droner_infinity.png'
+import upload_droner_infinity from '../../resource/media/empties/upload_Img_btn.png'
 import '../../pages/farmer/Style.css'
 import { DronerFinityDatasource } from '../../datasource/DronerFinityDatasource'
 import Select from 'react-select'
@@ -52,8 +52,6 @@ const ModalDronerInfinity: React.FC<ModalDronerInfinityProps> = ({
     SearchDronerByIdEntity_INIT,
   )
   const [dropdownVisible, setDropdownVisible] = useState(false)
-
-  const navigate = useNavigate()
   const [profile] = useLocalStorage('profile', [])
   const [createImg, setCreateImg] = useState<UploadImageEntity>(UploadImageEntity_INTI)
   const [editCreateImg, setEditCreateImg] = useState<UploadImageEntity>(UploadImageEntity_INTI)
@@ -397,7 +395,7 @@ const ModalDronerInfinity: React.FC<ModalDronerInfinityProps> = ({
               </div>
               <div className='col-lg' style={{ textAlign: 'end' }}>
                 <u style={{ color: color.Success, fontWeight: 500, cursor: 'pointer' }}>
-                  ดาวน์โหลดสัญญาณ
+                  ดาวน์โหลดสัญญา
                 </u>
               </div>
             </div>