+++ src/pages/task/newTask/AddNewTask.tsx
@@ -94,7 +94,6 @@ const AddNewTask = () => {
     useState<FarmerPlotEntity>(FarmerPlotEntity_INIT)
   const [selectionType] = useState<RowSelectionType>(queryString[1])
   const [checkSelectPlot, setCheckSelectPlot] = useState<any>('error')
-
   const [otherSpray, setOtherSpray] = useState<any>()
   const [cropSelected, setCropSelected] = useState<any>('')
   const [periodSpray, setPeriodSpray] = useState<CropPurposeSprayEntity>()
@@ -345,6 +344,11 @@ const AddNewTask = () => {
   const handleComment = (e: any) => {
     const d = Map(createNewTask).set('comment', e.target.value)
     setCreateNewTask(d.toJS())
+  }
+  const handlePreparationRemark = (e: any) => {
+    const d = Map(createNewTask).set('preparationRemark', e.target.value)
+    setCreateNewTask(d.toJS())
+    checkValidateStep(d.toJS(), current)
   }
   const handleDateAppointment = (e: any) => {
     setDateAppointment(moment(new Date(e)).format(dateCreateFormat))
@@ -771,15 +775,27 @@ const AddNewTask = () => {
                 <Radio value='นักบินโดรนเตรียมให้'>นักบินโดรนเตรียมให้</Radio>
               </Space>
             </Radio.Group>
-          </div>
-          <div className='form-group'>
-            <label>หมายเหตุ</label>
-            <TextArea
-              placeholder='ระบุหมายเหตุเพื่อแจ้งนักบินโดรน เช่น เกษตรกรจะเตรียมยาให้, ฝากนักบินเลือกยาราคาไม่แพงมาให้หน่อย เป็นต้น'
-              disabled={current === 2 || checkSelectPlot === 'error'}
-              onChange={handleComment}
-              defaultValue={createNewTask.comment}
-            />
+            {createNewTask.preparationBy === 'นักบินโดรนเตรียมให้' ? (
+              <div className='pt-3'>
+                <TextArea
+                  style={{ width: '530px', height: '80px', left: '4%' }}
+                  placeholder='(บังคับ) ระบุชื่อยา/ปุ๋ย และจำนวนที่ใช้'
+                  onChange={handlePreparationRemark}
+                  defaultValue={createNewTask.preparationRemark}
+                />
+              </div>
+            ) : null}
+
+            <div className='form-group pt-3'>
+              <label>หมายเหตุ</label>
+              <TextArea
+                style={{ left: 20, height: '80px' }}
+                placeholder='ระบุหมายเหตุเพื่อแจ้งนักบินโดรน เช่น เกษตรกรจะเตรียมยาให้, ฝากนักบินเลือกยาราคาไม่แพงมาให้หน่อย เป็นต้น'
+                disabled={current === 2 || checkSelectPlot === 'error'}
+                onChange={handleComment}
+                defaultValue={createNewTask.comment}
+              />
+            </div>
           </div>
         </Form>
       </div>
@@ -1353,7 +1369,6 @@ const AddNewTask = () => {
         data?.farmerId,
         data?.farmerPlotId,
         data?.purposeSprayId,
-        data?.preparationBy,
         data.farmAreaAmount,
       ].includes('')
       const checkEmptyNumber = ![data.price, data.unitPrice, data.farmAreaAmount].includes(0)
@@ -1366,11 +1381,18 @@ const AddNewTask = () => {
       }
       const checkOtherSpray = otherSpray && otherSpray.trim().length !== 1
       const checkDateTime = ![dateAppointment, timeAppointment].includes('')
+      const checkPreparationBy =
+        data.preparationBy === 'นักบินโดรนเตรียมให้' && data.preparationRemark.trim().length !== 0
 
       if (
         checkEmptyArray && [data?.targetSpray][0]?.includes('อื่นๆ')
           ? checkOtherSpray
-          : checkEmptyArray && checkDateTime && checkEmptyNumber && defaultRai && checkEmptySting
+          : checkEmptyArray &&
+            checkDateTime &&
+            checkEmptyNumber &&
+            defaultRai &&
+            checkEmptySting &&
+            checkPreparationBy
       ) {
         setDisableBtn(false)
       } else {